Replaced single ampersands in `npm` commands.
The `npm run android` and `npm run build:android` scripts were using single `&` to delimit their subcommands.  This actually runs the commands asynchronously, which means either the `build_hybrid.js` script _or_ the `cordova` command might execute first.

I think the intention here was to first run the `build_hybrid.js` script then run `cordova`, so I changed this to a `&&` which waits for the successful completion of the previous command to start the next one.